Move VegaWrapper to Common folder, use also in Markdown
