+++ client/src/components/Common/VegaWrapper.vue
