+++ client/src/components/Markdown/Sections/MarkdownVega.vue
@@ -1,0 +1,17 @@
+<script setup lang="ts">
+const VegaWrapper = () => import("@/components/Common/VegaWrapper.vue");
+import { computed } from "vue";
+
+const props = defineProps<{
+    content: string;
+}>();
+
+const spec = computed(() => ({
+    ...JSON.parse(props.content),
+    width: "container",
+}));
+</script>
+
+<template>
+    <VegaWrapper :spec="spec" />
+</template>
